# apispec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Specification",
    "description": "Dokumentasi API lengkap untuk backend service.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/emolog": {
      "get": {
        "tags": ["Emolog"],
        "summary": "Mendapatkan riwayat emolog pengguna",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Berhasil mengambil riwayat emolog" }
        }
      },
      "post": {
        "tags": ["Emolog"],
        "summary": "Mencatat entri emolog baru",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": { "type": "string", "example": "Hari ini saya merasa sangat senang sekali!" },
                  "subdistrict_id": { "type": "integer", "example": 1 }
                },
                "required": ["text", "subdistrict_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Emolog berhasil dicatat" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}